% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_site}
\alias{build_site}
\title{Betty Build}
\usage{
build_site(remote, dest = ".", deploy_org = "ropensci-docs",
  deploy_url = "https://docs.ropensci.org")
}
\arguments{
\item{remote}{url of the git remote}

\item{dest}{path of volume to save docs and src output}

\item{deploy_org}{name of github organization to publish gh-pages repo}

\item{deploy_url}{optional base domain name under which sites will be hosted.
Defaults to `https://{deploy_org}.github.io`.}
}
\description{
Builds pkgdown site and source package from  a git remote and optionally
publish it directly to \url{https://docs.ropensci.org}.
}
\examples{
\dontrun{
build_site('https://github.com/ropensci/magick')
}
}
