% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_site}
\alias{deploy_site}
\title{Betty Deploy}
\usage{
deploy_site(path, deploy_org, deploy_url = NULL)
}
\arguments{
\item{path}{local path of the website root directory}

\item{deploy_org}{name of github organization to publish gh-pages repo}

\item{deploy_url}{optional base domain name under which sites will be hosted.
Defaults to `https://{deploy_org}.github.io`.}
}
\description{
Publish a local site to our ropensci-docs github org. Make sure you have
a valid `GITHUB_PAT` token in your environment to use this non-interactively.
}
