% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\alias{build_site}
\title{Build}
\usage{
build_site(remote, dest = ".", deploy = FALSE)
}
\arguments{
\item{remote}{url of the git remote}

\item{dest}{path of volume to save docs and src output}

\item{deploy}{publish website on docs.ropensci.org}
}
\description{
Builds pkgdown site and source package from  a git remote and optionally
publish it directly to \url{https://docs.ropensci.org}.
}
