% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_site}
\alias{deploy_site}
\alias{deploy_all_sites}
\title{Betty Deploy}
\usage{
deploy_site(path, deploy_org)

deploy_all_sites(doc_root, deploy_org = "ropensci-docs")
}
\arguments{
\item{path}{local path of the website root directory}

\item{deploy_org}{name of github organization to publish gh-pages repo
Defaults to `https://{deploy_org}.github.io/pkg`.}

\item{doc_root}{directory in which your websites are stored}
}
\description{
Publish a local site or all sites in a given directory to a github org. Set
a valid `GITHUB_PAT` token in your environment to use this non-interactively.
}
