% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\alias{deploy_site}
\title{Deploy}
\usage{
deploy_site(path = ".")
}
\arguments{
\item{path}{local path of the website}
}
\description{
Publish a local site to our ropensci-docs github org. You need a
valid `GITHUB_PAT` token in your environment to deploy.
}
