% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{sync_ropensci_jenkins}
\alias{sync_ropensci_jenkins}
\alias{sync_ropensci_docs}
\alias{config_template}
\alias{view_template}
\title{Sync rOpenSci with Jenkins}
\usage{
sync_ropensci_jenkins(update_jobs = FALSE, remove_jobs = TRUE,
  update_views = TRUE)

sync_ropensci_docs()

config_template(git_url)

view_template(view_jobs)
}
\arguments{
\item{update_jobs}{update the xml config of existing repos.}

\item{remove_jobs}{delete jobs that are no longer in the registry}

\item{update_views}{update the views (per-author package list)}

\item{git_url}{HTTPS git url of the target repository}

\item{view_jobs}{Character vector with jobs to add to this view}
}
\description{
Checks the rOpenSci registery and add/delete corresponding projects on Jenkins
and the ropensci-docs repositories on Github.
}
